+++ platform/darwin/src/MGLRasterTileSource_Private.h
@@ -1,8 +1,10 @@
 #import "MGLRasterTileSource.h"
 
 #include <memory>
+#include <mbgl/util/variant.hpp>
 
 namespace mbgl {
+    class Tileset;
     namespace style {
         class RasterSource;
     }
@@ -14,7 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property (nonatomic, readonly, nullable) mbgl::style::RasterSource *rawSource;
 
-- (std::unique_ptr<mbgl::style::RasterSource>)pendingSourceWithIdentifier:(NSString *)identifier configurationURL:(NSURL *)configurationURL tileSize:(CGFloat)tileSize;
+- (std::unique_ptr<mbgl::style::RasterSource>)pendingSourceWithIdentifier:(NSString *)identifier urlOrTileset:(mbgl::variant<std::string, mbgl::Tileset>)urlOrTileset tileSize:(uint16_t)tileSize;
 
 @end