+++ src/mbgl/renderer/buckets/line_bucket.hpp
@@ -26,16 +26,16 @@ public:
                const uint32_t overscaling);
     ~LineBucket() override;
 
-    void addFeature(const GeometryTileFeature&,
-                    const GeometryCollection&,
-                    const mbgl::ImagePositions& patternPositions,
-                    const PatternLayerMap&) override;
+    void addFeature(const GeometryTileFeature&, const GeometryCollection&, const mbgl::ImagePositions& patternPositions,
+                    const PatternLayerMap&, std::size_t) override;
 
     bool hasData() const override;
 
     void upload(gfx::UploadPass&) override;
 
     float getQueryRadius(const RenderLayer&) const override;
+
+    void update(const FeatureStates&, const GeometryTileLayer&, const std::string&, const ImagePositions&) override;
 
     PossiblyEvaluatedLayoutProperties layout;